requirements.txt
django==4.0.0
channels==3.0.5
djangorestframework==3.14.0
channels-redis==4.1.0
pyjwt==2.8.0
django-cors-headers==4.3.1
